backend/FitFever/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
cssselect2==0.7.0
Django==4.2.7
django-cors-headers==4.3.1
django-cron==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
fonttools==4.53.0
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.6
msgpack==1.0.8
packaging==24.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.10.0
PyJWT==2.8.0
pyparsing==3.1.2
pyphen==0.15.0
python-decouple==3.8
python-http-client==3.3.7
python-magic==0.4.27
pytz==2024.1
requests==2.31.0
rsa==4.9
sendgrid==6.11.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tinycss2==1.3.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==62.2
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
